app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
PyJWT
pwdlib[argon2]
email-validator
tomlkit
python-multipart
pdfplumber
python-docx
langchain-community
langchain-core
langchain-text-splitters
chromadb
sentence-transformers
transformers
torch
numpy
scikit-learn